+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [6.4.2] - 2021-01-03
+
+### Fixed
+
+- N3DR does not fail if parameter is empty and not populated in
+  `~/.n3dr/config.yml` reported by
+  [der-eismann](https://github.com/der-eismann).
+
 ## [6.4.1] - 2021-01-03
 
 ### Changed
@@ -604,7 +612,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Download all artifacts from a certain Nexus3 repository.
 
-[Unreleased]: https://github.com/030/n3dr/compare/6.4.1...HEAD
+[Unreleased]: https://github.com/030/n3dr/compare/6.4.2...HEAD
+[6.4.2]: https://github.com/030/n3dr/compare/6.4.1...6.4.2
 [6.4.1]: https://github.com/030/n3dr/compare/6.4.0...6.4.1
 [6.4.0]: https://github.com/030/n3dr/compare/6.3.0...6.4.0
 [6.3.0]: https://github.com/030/n3dr/compare/6.2.0...6.3.0